@@ -11,24 +11,25 @@
     "cdk": "cdk"
   },
   "devDependencies": {
-    "@aws-cdk/assert": "1.96.0",
-    "@types/jest": "^26.0.22",
-    "@types/node": "*",
-    "aws-cdk": "1.96.0",
-    "jest": "^26.6.3",
-    "ts-jest": "^26.5.4",
-    "ts-node": "^9.1.1",
-    "typescript": "~4.2.3"
+    "@aws-cdk/assert": "^2.68.0",
+    "@types/jest": "^29.5.5",
+    "@types/node": "^20.8.6",
+    "aws-cdk": "^2.101.0",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   },
   "dependencies": {
-    "@aws-cdk/aws-apigateway": "1.96.0",
-    "@aws-cdk/aws-dynamodb": "1.96.0",
-    "@aws-cdk/aws-lambda": "1.96.0",
-    "@aws-cdk/aws-sns": "1.96.0",
-    "@aws-cdk/aws-sns-subscriptions": "1.96.0",
-    "@aws-cdk/aws-sqs": "1.96.0",
-    "@aws-cdk/core": "1.96.0",
-    "aws-sdk": "^2.1167.0",
-    "cdk-dynamo-table-viewer": "^3.1.2"
+    "@aws-cdk/aws-apigateway": "^1.204.0",
+    "@aws-cdk/aws-dynamodb": "^1.204.0",
+    "@aws-cdk/aws-lambda": "*",
+    "@aws-cdk/aws-sns": "*",
+    "@aws-cdk/aws-sns-subscriptions": "^1.204.0",
+    "@aws-cdk/aws-sqs": "*",
+    "@aws-cdk/core": "*",
+    "@aws-sdk/client-dynamodb": "^3.428.0",
+    "aws-sdk": "^2.1473.0",
+    "cdk-dynamo-table-viewer": "^0.2.487"
   }
 }
