@@ -1,4 +1,4 @@
-const { DynamoDB } = require("aws-sdk");
+import { DynamoDB } from "@aws-sdk/client-dynamodb";
 
 exports.handler = async function (event) {
   console.log("request:", JSON.stringify(event, undefined, 2));
